Add global Event Token storage methods
Add global Event Token storage for use in attached dependency
properties.  This allows us to store our event tokens on the object so
they are easy to find later when we want to remove them, but when we
don't have an associated class or other place to easily put the data
(such as when writing attached dependency properties).